+++ app/src/main/java/com/example/hangit/ranking/user.kt
@@ -2,5 +2,5 @@ package com.example.hangit.ranking
 
 data class User(
     val name: String = "",
-    val score: Int = 0
+    val score: Long = 0
 )
